second changes

+++ AutoPXC.sh
@@ -6,12 +6,12 @@ nodeip=`ifconfig | grep 'inet' | cut -d: -f2 | awk '{ print $2}' | grep -v '^$' 
 
 echo "USER:$USER  TIME:`date +%Y-%m-%d\ %H:%M:%S` HOST:$HOSTNAME IP:$nodeip"
 
-echo "$nodeip $HOSTNAME" >> /etc/hosts
+echo "$nodeip $HOSTNAME" >> /etc/hosts #ke xuan
 
 datadir="" #mysql basedir
 
-ErrMSG="directory is exsit."
-DstMSG="pls input mysqldata directory."
+ErrMSG="Directory is exsit."
+DstMSG="Enter the path of mysqldata directory."
 
 function inputDstPath(){
 
@@ -20,18 +20,93 @@ function inputDstPath(){
 	if [ -d $datadir ];then
 	echo "$ErrMSG"	
 	useradd -s /sbin/nologin mysql >/dev/null &>1
-	chown mysql.mysql -R $datadir && echo `ll $datadir`
+	chown mysql.mysql -R $datadir
+	ll $datadir
 	else
 	mkdir -p $datadir
 	useradd -s /sbin/nologin mysql >/dev/null &>1
-	chown mysql.mysql -R $datadir && echo `ll $datadir`
+	chown mysql.mysql -R $datadir
+	ll $datadir
 	fi
 }
 
 function makecnf(){
 
-read -p "input all ip of machines ,like 192.168.xxx.xxx,192.168.xxx.xxx,:" allip
-read -p "input server-id like 1,2,3...,make sure every machine difference:" sid
+#read -p "Enter the IP of all the nodes,like 192.168.xxx.xxx,192.168.xxx.xxx,192.168.xxx.xxx:" allip
+#read -p "input server-id like 1,2,3...,make sure every machine difference:" sid
+
+sid=`date +%s%N | cut -c17-19`
+
+cpun=`cat /proc/cpuinfo| grep "processor"| wc -l`
+cp /etc/my.cnf my.cnf.bak > /dev/null &>1
+
+echo "
+[mysqld]
+datadir=$datadir
+socket = /tmp/mysql.sock
+pid-file=$datadir/mysql.pid
+character_set_server = utf8
+max_connections = 3000
+back_log= 3000
+skip-name-resolve
+sync_binlog=0
+innodb_flush_log_at_trx_commit=1
+server-id = $sid
+default_storage_engine=Innodb
+innodb_autoinc_lock_mode=2
+binlog_format=row
+wsrep_cluster_name=pxc_zs
+wsrep_slave_threads=$cpun #开启的复制线程数，cpu核数*2
+wsrep_cluster_address=gcomm://$nodeip,$node02,$node03
+wsrep_node_address=$nodeip
+wsrep_provider=/usr/local/mysql/lib/libgalera_smm.so
+wsrep_sst_method=xtrabackup-v2 
+wsrep_sst_auth=sst:zs " > /etc/my.cnf
+
+}
+
+function makecnf2(){
+
+#read -p "Enter the IP of all the nodes,like 192.168.xxx.xxx,192.168.xxx.xxx,192.168.xxx.xxx:" allip
+#read -p "input server-id like 1,2,3...,make sure every machine difference:" sid
+
+sid=`date +%s%N | cut -c17-19`
+
+cpun=`cat /proc/cpuinfo| grep "processor"| wc -l`
+cp /etc/my.cnf my.cnf.bak > /dev/null &>1
+
+echo "
+[mysqld]
+datadir=$datadir
+socket = /tmp/mysql.sock
+pid-file=$datadir/mysql.pid
+character_set_server = utf8
+max_connections = 3000
+back_log= 3000
+skip-name-resolve
+sync_binlog=0
+innodb_flush_log_at_trx_commit=1
+server-id = $sid
+default_storage_engine=Innodb
+innodb_autoinc_lock_mode=2
+binlog_format=row
+wsrep_cluster_name=pxc_zs
+wsrep_slave_threads=$cpun #开启的复制线程数，cpu核数*2
+wsrep_cluster_address=gcomm://$node01ip,$nodeip,$node03
+wsrep_node_address=$nodeip
+wsrep_provider=/usr/local/mysql/lib/libgalera_smm.so
+wsrep_sst_method=xtrabackup-v2 
+wsrep_sst_auth=sst:zs " > /etc/my.cnf
+
+}
+
+function makeall(){
+
+read -p "Enter the IP of all the nodes,like 192.168.xxx.xxx,192.168.xxx.xxx,192.168.xxx.xxx :" allip
+#read -p "input server-id like 1,2,3...,make sure every machine difference:" sid
+
+sid=`date +%s%N | cut -c17-19`
+
 cpun=`cat /proc/cpuinfo| grep "processor"| wc -l`
 cp /etc/my.cnf my.cnf.bak > /dev/null &>1
 
@@ -60,6 +135,44 @@ wsrep_sst_auth=sst:zs " > /etc/my.cnf
 
 }
 
+
+function makecnf3(){
+
+#read -p "Enter the IP of all the nodes,like 192.168.xxx.xxx,192.168.xxx.xxx,192.168.xxx.xxx:" allip
+#read -p "input server-id like 1,2,3...,make sure every machine difference:" sid
+
+sid=`date +%s%N | cut -c17-19`
+
+cpun=`cat /proc/cpuinfo| grep "processor"| wc -l`
+cp /etc/my.cnf my.cnf.bak > /dev/null &>1
+
+echo "
+[mysqld]
+datadir=$datadir
+socket = /tmp/mysql.sock
+pid-file=$datadir/mysql.pid
+character_set_server = utf8
+max_connections = 3000
+back_log= 3000
+skip-name-resolve
+sync_binlog=0
+innodb_flush_log_at_trx_commit=1
+server-id = $sid
+default_storage_engine=Innodb
+innodb_autoinc_lock_mode=2
+binlog_format=row
+wsrep_cluster_name=pxc_zs
+wsrep_slave_threads=$cpun #开启的复制线程数，cpu核数*2
+wsrep_cluster_address=gcomm://$node01ip,$node02,$nodeip
+wsrep_node_address=$nodeip
+wsrep_provider=/usr/local/mysql/lib/libgalera_smm.so
+wsrep_sst_method=xtrabackup-v2 
+wsrep_sst_auth=sst:zs " > /etc/my.cnf
+
+}
+
+
+
 function download(){
 
 echo "downloading..."
@@ -79,9 +192,9 @@ fi
 
 function copy(){
 
-read -p "input node01's ip address:" node1ip
+read -p "Enter the IP of node01 :" node01ip
 
-scp root@$node1ip:/root/percona-xtrabackup-2.4.6-Linux-x86_64.tar.gz root@$node1ip:/root/Percona-XtraDB-Cluster-5.6.26-rel74.0-25.12.1.Linux.x86_64.tar.gz /root && echo "copy complete"
+scp root@$node01ip:/root/percona-xtrabackup-2.4.6-Linux-x86_64.tar.gz root@$node01ip:/root/Percona-XtraDB-Cluster-5.6.26-rel74.0-25.12.1.Linux.x86_64.tar.gz /root && echo "copy complete"
 
 }
 
@@ -96,7 +209,7 @@ tar xvf /root/Percona-XtraDB-Cluster-5.6.26-rel74.0-25.12.1.Linux.x86_64.tar.gz 
 chown mysql.mysql -R mysql
 cp percona-xtrabackup-2.4.6-Linux-x86_64/bin/* mysql/bin/
 
-yum remove mariadb-* -y
+yum remove mariadb-* -y >/dev/null &>1
 
 yum install perl-IO-Socket-SSL.noarch perl-DBD-MySQL.x86_64 perl-Time-HiRes openssl openssl-devel socat -y
 
@@ -107,7 +220,7 @@ ln -s /usr/lib64/libssl.so.10 /lib64/libssl.so.6 >/dev/null &>1
 }
 
 
-installdb(){
+function installdb(){
 
 echo "export PATH=$PATH:/usr/local/mysql/bin" > /etc/profile.d/mysql.sh && source /etc/profile.d/mysql.sh
 
@@ -118,12 +231,39 @@ echo "export PATH=$PATH:/usr/local/mysql/bin" > /etc/profile.d/mysql.sh && sourc
 function mysqlv(){
 
 if [ -f /etc/init.d/mysql ];then
-	/etc/init.d/mysql bootstrap-pxc	&& mysql -v
+	/etc/init.d/mysql bootstrap-pxc
+	pxcstat=`/etc/init.d/mysql bootstrap-pxc`
+	if [ "$pxcstat" == "*SUCCESS*" ];then
+	mysql -v
 	delete from mysql.user where user!='root' or host!='localhost';
 	grant all privileges on *.* to 'sst'@'%' identified by 'zs';
 	grant all privileges on *.* to 'sst'@'localhost' identified by 'zs';
 	flush privileges;
 	quit;
+	else
+	break
+	fi
+else
+echo "/etc/init.d/mysql is not exsit,quit."
+break
+fi
+}
+
+function mysql2(){
+
+if [ -f /etc/init.d/mysql ];then
+	/etc/init.d/mysql start
+	pxcstat=`/etc/init.d/mysql start`
+	if [ "$pxcstat" == "*SUCCESS*" ];then
+	mysql -v
+	delete from mysql.user where user!='root' or host!='localhost';
+	grant all privileges on *.* to 'sst'@'%' identified by 'zs';
+	grant all privileges on *.* to 'sst'@'localhost' identified by 'zs';
+	flush privileges;
+	quit;
+	else
+	break
+	fi
 else
 echo "/etc/init.d/mysql is not exsit,quit."
 break
@@ -134,7 +274,7 @@ fi
 while true
 do
 
-echo "***************************************"
+echo "*************************************************************"
 cat << EOF
     1.node01
     2.node02
@@ -142,16 +282,22 @@ cat << EOF
     4.quit
     *iuput the NO. of machine*
     When finish all machines,login mysql of node01, 
-    use "show status like 'wsrep%';" check Cluster state,
+    Use "show status like 'wsrep%';" check Cluster state,
     If state is ok, use "mysql_secure_installation" secure the Cluster.
 EOF
-echo "***************************************"
+echo "*************************************************************"
 
 read -p ":" op
 
 case $op in
 	1)
 	echo "master-node install"
+	read -p "enter the IP of node02 :" node02
+	read -p "enter the hostname of node02 :" hostname02
+	echo "$node02 $hostname02" >> /etc/hosts
+	read -p "enter the IP of node03 :" node03
+	read -p "enter the hostname of node03 :" hostname03
+	echo "$node03 $hostname03" >> /etc/hosts
 	inputDstPath
 	makecnf
 	download
@@ -159,19 +305,49 @@ case $op in
 	installdb
 	mysqlv
 	break
+	;;       
+	2)
+	echo "slave-node02 install"
+	inputDstPath
+	copy
+	read -p "enter the hostname of node01 :" hostname01
+	echo "$node01ip $hostname01" >> /etc/hosts
+	read -p "enter the IP of node03 :" node03
+	read -p "enter the hostname of node03 :" hostname03
+	echo "$node03 $hostname03" >> /etc/hosts
+	makecnf2
+	tarfile
+	installdb
+	mysql2
+	break
 	;;
-        4|quit)
-	echo "Exit..."
+	3)
+	echo "slave-node03 install"
+	inputDstPath
+	copy
+	read -p "enter the hostname of node01 :" hostname01
+	echo "$node01ip $hostname01" >> /etc/hosts
+	read -p "enter the hostname of node02 :" hostname02	
+	read -p "enter the IP of node02 :" node02
+	echo "$node02 $hostname02" >> /etc/hosts
+	makecnf3
+	tarfile
+	installdb
+	mysql2
 	break
 	;;
 	*)
 	echo "slave-node install"
 	inputDstPath
-	makecnf
 	copy
+	makeall
 	tarfile
 	installdb
-	mysqlv
+	mysql2
+	break
+	;;
+	4|quit)
+	echo "Exit..."
 	break
 	;;
 esac
